Split acquisition into synthetic and redpitaya packages
