--- self_interferometry/acquisition/__init__.py
@@ -1,5 +1,0 @@
-"""Data acquisition and simulation modules."""
-
-from . import redpitaya, simulations
-
-__all__ = ['redpitaya', 'simulations']
