--- self_interferometry/acquisition/redpitaya/__init__.py
@@ -1,7 +1,0 @@
-"""Red Pitaya hardware interface module."""
-
-from .manager import RedPitayaManager
-from .redpitaya_config import RedPitayaConfig
-from .scpi import SCPI
-
-__all__ = ['SCPI', 'RedPitayaConfig', 'RedPitayaManager']
