+++ self_interferometry/analysis/datasets.py
@@ -8,8 +8,8 @@ import numpy as np
 import torch
 from torch.utils.data import DataLoader, Dataset
 
-from self_interferometry.acquisition.redpitaya.redpitaya_config import RedPitayaConfig
-from self_interferometry.acquisition.simulations.coil_driver import CoilDriver
+from self_interferometry.redpitaya.redpitaya_config import RedPitayaConfig
+from self_interferometry.synthetic.coil_driver import CoilDriver
 
 logger = logging.getLogger(__name__)
