+++ self_interferometry/analysis/inspect_data.py
@@ -8,8 +8,8 @@ import matplotlib.pyplot as plt
 import numpy as np
 from torch.utils.data import DataLoader
 
-from self_interferometry.acquisition.redpitaya.redpitaya_config import RedPitayaConfig
 from self_interferometry.analysis.datasets import VelocityDataset
+from self_interferometry.redpitaya.redpitaya_config import RedPitayaConfig
 
 logger = logging.getLogger(__name__)
