+++ self_interferometry/main.py
@@ -5,9 +5,10 @@ import time
 from pathlib import Path
 
 import lightning as lightning_module
-from acquisition.redpitaya.manager import RedPitayaManager
 from analysis.generate_data import generate_dataset_from_rp
 from analysis.training_interface import TrainingConfig, TrainingInterface
+
+from self_interferometry.redpitaya.manager import RedPitayaManager
 
 
 def setup_logging(verbosity: str) -> None:
