+++ self_interferometry/synthetic/__init__.py
