+++ self_interferometry/synthetic/calib_params.py
