+++ self_interferometry/synthetic/interferometers.py
