+++ self_interferometry/synthetic/waveform.py
